feat: implement centralized UI handling and improved animal interaction logic
Added MainGameUIHandler to manage info panel updates and toggling
Moved UI text updates and show/hide logic from Animal scripts to handler
Added raycast-based click detection to close panel when clicking outside animals
Improved Animal subclasses (Cat, Dog, Chicken) to use shared UI handler
Cleaned up GameManager, SaveSystem, and scene setup
Updated MainGame scene layout and NavMesh configuration

+++ ProgrammingTheoryProject/Assets/Scripts/Entities/Helper/AgentAnimatorBridge.cs
@@ -27,6 +27,6 @@ public class AgentAnimatorBridge : MonoBehaviour
         int state = 0; // idle
         if (s > runThreshold) state = 2;
         else if (s > walkThreshold) state = 1;
-        anim.SetInteger(stateParam, state);
+        anim.SetFloat(stateParam, (float)state);
     }
 }
